Tying up some loose ends
- Added support for IComparable in the AssertEqualityComparer
- Added XML comments
- Moved classes to their own files
